DistModel {dismo}	R Documentation

Class "DistModel"

Description

Parent class for a number of distribution models defined in the dismo package (those created by bioclim, domain, maxent and mahal). This is a virtual Class, no objects may be direclty created from it.

Slots

presence:

presence data used

absence:

absence or background data used

hasabsence:

Logical indicating whether there is any absence data

Author(s)

Robert J. Hijmans

[Package dismo version 1.1-1 Index]