calc.deviance {dismo}	R Documentation

Calculate deviance

Description

Function to calculate deviance given two vectors of observed and predicted values. Requires a family argument which is set to binomial by default

Usage

calc.deviance(obs, pred, weights = rep(1,length(obs)), 
                family="binomial", calc.mean = TRUE)

Arguments

obs
a vector with observed values

pred
a vector with predicted values that correspond the the values in obs

weights
a vector of weight values

family
One of "binomial", "bernoulli", "poisson", "laplace", or "gaussian"

calc.mean
Logical. If TRUE, the mean deviance is returned

Author(s)

John R. Leathwick and Jane Elith

[Package dismo version 1.1-1 Index]