dcEvaluate {dismo}	R Documentation

Evaluate by distance class

Description

Evaluate a model for intervals of distances to the nearest point in a reference dataset for presence data and a sample of the absence data selected to have a low spatial sorting bias (obtained with pwdSample).

Usage

dcEvaluate(p, a, reference, lonlat=TRUE, binsize=15, predp, preda, model, 
               predictors, fun=predict) 

Arguments

p
two column matrix (x, y) or (longitude/latitude) or SpatialPoints object, for point locations

a
two column matrix (x, y) or (longitude/latitude) or SpatialPoints object, for point locations

reference
as above for reference point locations to which distances are computed

lonlat
Logical. Use TRUE if the coordinates are spherical (in degrees), and use FALSE if they are planar

binsize
postive integer. How many presence points in each distance bin?

predp
p

preda
a

model
m

predictors
pr

fun
function

Value

list with Evaluation objects

Author(s)

Robert J. Hijmans

See Also

pwdSample, ssb

[Package dismo version 1.0-15 Index]