density {dismo}	R Documentation

density

Description

Create a density plots of presence and absence data

Value

A density plot. Presence data are in red, and absence data (if available) are in blue.

Methods

density(x, ...)

x 	 Object of class 'ModelEvaluation' or of a class that inherits from 'DistModel, (such as 'MaxEnt', 'Bioclim')

... 	 Additional arguments that can be passed to plot

Author(s)

Robert J. Hijmans

See Also

evaluate

[Package dismo version 1.1-1 Index]