evaluate {dismo}	R Documentation

Model evaluation

Description

Cross-validation of models with presence/absence data. Given a vector of presence and a vector of absence values (or a model and presence and absence points and predictors), confusion matrices are computed (for varying thresholds), and model evaluation statistics are computed for each confusion matrix / threshold. See the description of class ModelEvaluation-class for more info.

Usage

evaluate(p, a, model, x, tr, ...)

Arguments

p
presence points (x and y coordinates or SpatialPoints* object).

Or, if x is missing, values at presence points

Or, a matrix with values to compute predictions for

a
absence points (x and y coordinates or SpatialPoints* object).

Or, if x is missing, values at presence points.

Or, a matrix with values to compute predictions for

model
any fitted model, including objects inherting from 'DistModel'; not used when x is missing

x
Optional. Predictor variables (object of class Raster*). If present, p and a are interpreted as (spatial) points

tr
Optional. a vector of threshold values to use for computing the confusion matrices

...
Additional arguments for the predict function

Value

An object of ModelEvaluation-class

Author(s)

Robert J. Hijmans

References

Fielding, A.H. and J.F. Bell, 1997. A review of methods for the assessment of prediction errors in conservation presence/absence models. Environmental Conservation 24:38-49

See Also

threshold

Examples

## See ?maxent for an example with real data.
# this is a contrived example:
# p has the predicted values for 50 known cases (locations) 
# with presence of the phenomenon (species)
p <- rnorm(50, mean=0.7, sd=0.3)
# b has the predicted values for 50 background locations (or absence)
a <- rnorm(50, mean=0.4, sd=0.4)
e <- evaluate(p=p, a=a)

threshold(e)

plot(e, 'ROC')
plot(e, 'TPR')
boxplot(e)
density(e)

str(e)

[Package dismo version 1.1-1 Index]