evaluateROCR {dismo}	R Documentation

Model testing with the ROCR package

Description

Preparing data for model testing with the ROCR package.

Usage

evaluateROCR(model, p, a, x)

Arguments

model
any fitted model, including objects inherting from 'DistModel'

p
presence points (x and y coordinates or SpatialPoints* object).

Or, if x is missing, values at presence points

Or, a matrix with values to compute predictions for

a
absence points (x and y coordinates or SpatialPoints* object).

Or, if x is missing, values at presence points.

Or, a matrix with values to compute predictions for

x
optional. predictor variables, if present, p and a are considered

Value

An object of class "prediction" (defined in the ROCR package)

Author(s)

Robert J. Hijmans

[Package dismo version 1.1-1 Index]