Anguilla data {dismo}	R Documentation

Anguilla australis distribution data

Description

A number of sites with presence or absence of the short-finned eel (Anguilla australis) in New Zealand, and environmental data at these sites; and gridded data of the environmental variables for the study area.

type 	 variable name 	 values 	 mean and range

Reach 	 LocSed 	 weighted average of proportional cover of bed sediment 	 1=mud, 2=sand, 3=fine gravel; 4=coarse gravel; 5=cobble; 6=boulder; 7=bedrock; 3.77, 1 to 7

Segment 	 SegSumT 	 Summer air temperature (degrees C) 	 16.3, 8.9 to 19.8

SegTSeas 	 Winter air temperature (degrees C), normalised with respect to SegJanT 	 0.36, -4.2 to 4.1

SegLowFlow 	 segment low flow (m3/sec), fourth root transformed 	 1.092, 1.0 to 4.09

Downstream 	 DSDist 	 distance to coast (km) 	 74, 0.03 to 433.4

DSDam 	 presence of known downstream obstructions, mostly dams 	 0.18, 0 or 1

DSMaxSlope 	 maximum downstream slope (degrees) 	 3.1, 0 to 29.7

Upstream / catchment 	 USAvgT 	 average temperature in catchment (deg C) compared to segment, normalised with respect to SegJanT 	 -0.38, -7.7 to 2.2

USRainDays 	 days/month with rain greater than 25 mm 	 1.22, 0.21 to 3.30

USSlope 	 average slope in the upstream catchment (degrees) 	 14.3, 0 to 41.0

USNative 	 area with indigenous forest (proportion) 	 0.57, 0 to 1

Fishing method 	 fishing method in five classes: electric, net, spot, trap & mixture 	 NA

Usage

data(Anguilla_train)
data(Anguilla_test)
data(Anguilla_grids)

Author(s)

John R. Leathwick and Jane Elith

References

Elith, J., J.R. Leathwick and T. Hastie, 2009. A working guide to boosted regression trees. Journal of Animal Ecology 77: 802-81

[Package dismo version 1.1-1 Index]