gbm.perspec {dismo}	R Documentation

gbm perspective plot

Description

Takes a gbm boosted regression tree object produced by gbm.step and plots a perspective plot showing predicted values for two predictors as specified by number using x and y. Values for all other variables are set at their mean by default but values can be specified by giving a list consisting of the variable name and its desired value, e.g., c(name1 = 12.2, name2 = 57.6)

Usage

gbm.perspec(gbm.object, x = 1, y = 2, pred.means = NULL, x.label = NULL, x.range = NULL,
 y.label = NULL, z.label = "fitted value", y.range = NULL, z.range = NULL, 
 leg.coords = NULL, ticktype = "detailed", theta = 55, phi = 40, smooth = "none",
 mask = FALSE, perspective = TRUE, ...)

Arguments

gbm.object
object of class gbm

x
the first variable to be plotted

y
the second variable to be plotted

pred.means
allows specification of values for other variables

x.label
allows manual specification of the x label

x.range
manual range specification for the x variable

y.label
and y label

z.label
default z label

y.range
and the y

z.range
allows control of the vertical axis

leg.coords
can specify coords (x, y) for legend

ticktype
specifiy detailed types - otherwise "simple"

theta
rotation

phi
and elevation

smooth
controls smoothing of the predicted surface

mask
controls masking using a sample intensity model

perspective
controls whether a contour or perspective plot is drawn

...
allows the passing of additional arguments to plotting routine useful options include shade, ltheta, lphi for controlling illumination and cex for controlling text size - cex.axis and cex.lab have no effect

Author(s)

John R. Leathwick and Jane Elith

References

Elith, J., J.R. Leathwick and T. Hastie, 2009. A working guide to boosted regression trees. Journal of Animal Ecology 77: 802-81

[Package dismo version 1.1-1 Index]