gbm.plot {dismo}	R Documentation

gbm plot

Description

Function to plot gbm response variables, with the option of adding a smooth representation of the response if requested additional options in this version allow for plotting on a common scale. Note hat fitted functions are centered by subtracting their mean.

Usage

gbm.plot(gbm.object, variable.no=0, smooth=FALSE, rug=TRUE, n.plots=length(pred.names),
  common.scale=TRUE, write.title=TRUE, y.label="fitted function", x.label=NULL, 
  show.contrib=TRUE, plot.layout=c(3, 4), ...)

Arguments

gbm.object
a gbm object - could be one from gbm.step

variable.no
the var to plot - if zero then plots all

smooth
Logical. If TRUE, a smoothed version of the fitted function is added

rug
Logical. If TRUE, a rug of deciles is plotted

n.plots
plot the first n most important preds

common.scale
Logical. If TRUE, a common scale is used on the y axis

write.title
Logical. If TRUE, the plot gets a title

y.label
the default y-axis label

x.label
the default x-axis label

show.contrib
Logical. If TRUE, the contribution is shown on the x axis

plot.layout
define the default layout for graphs on the page

...
other arguments to pass to the plotting useful options include cex.axis, cex.lab, etc.

Author(s)

John R. Leathwick and Jane Elith

References

Elith, J., J.R. Leathwick and T. Hastie, 2009. A working guide to boosted regression trees. Journal of Animal Ecology 77: 802-81

[Package dismo version 1.1-1 Index]