gbm.plot.fits {dismo}	R Documentation

gbm plot fitted values

Description

Plots the fitted values from a gbm object returned by any of the model fitting options. This can give a more reliable guide to the shape of the fitted surface than can be obtained from the individual functions, particularly when predictor variables are correlated and/or samples are unevenly distributed in environmental space. Allows masking out of absences to enable focus on sites with high predicted values.

Usage

gbm.plot.fits(gbm.object, v=0, mask.presence=FALSE, use.factor=FALSE )

Arguments

gbm.object
a gbm object

v
variable numbers to be plotted (if 0 then all are plotted)

mask.presence
Logical. If TRUE, the function only plots fitted values for presences

use.factor
Logical. If TRUE, forces to use quicker printing box and whisker plot

Author(s)

John R. Leathwick and Jane Elith

References

Elith, J., J.R. Leathwick and T. Hastie, 2009. A working guide to boosted regression trees. Journal of Animal Ecology 77: 802-81

[Package dismo version 1.1-1 Index]