geocode {dismo}	R Documentation

Georeferencing with Google

Description

A wrapper around the Google geocoding web-service. It returns 0 to n matches. It is important to be as precise as possible, e.g. always include the country in the locality description.

The purpose of using this function should be to display the locations on a map in a browser. You should check the Google terms of use http://code.google.com/apis/maps/terms.html to see if your usage of this function (and the underlying Google API) is permitted.

Usage

geocode(x, oneRecord=FALSE, extent=NULL, progress='', ...)

Arguments

x
A vector of locality descriptions

oneRecord
Logical. If TRUE a single record for each item in x is returned. If the API returned multiple records, the values of this record are computed by averaging the coordinates and taking the union of all bounding boxes

extent
An Extent object, or an object that can be coerced to one, to bias the search towards that region

progress
Character. Valid values are "" (no progress indicator), "text" or "window"

...
additional arguments (currently none implemeted)

Value

data.frame with the following fields:

originalPlace
the locality description as provided (in argument x)

interpretedPlace
the locality as interpreted by the Google API

lon
longitude

lat
latitude

lonmin
minimum longitude of the bounding box

lonmax
maximum longitude of the bounding box

latmin
minimum latitude of the bounding box

latmax
maximum latitude of the bounding box

uncertainty
distance from c(lon, lat) to the farthest corner of the bounding box

Note

It is important to compare fields originalPlace and interpretedPlace as the Google interpretation of a (perhaps vague) locality description can be very speculative

Author(s)

Robert J. Hijmans

Examples

## Not run: 
geocode(c('1600 Pennsylvania Ave NW, Washington DC', 'Luca, Italy', 'Kampala'))
geocode(c('San Jose', 'San Jose, Mexico'))
geocode(c('San Jose', 'San Jose, Mexico'), oneRecord=TRUE)

## End(Not run)

[Package dismo version 1.1-1 Index]