nicheEquivalency {dismo}	R Documentation

Niche equivalency

Description

Compute niche equivalency for two species following Warren et al. (2009). The statistic ranges from 0 to 1.

Usage

 
nicheEquivalency(sp1, sp2, predictors, n=99, model=maxent, verbose=TRUE, ...) 

Arguments

sp1
coordinates for species 1 (matrix with (x, y) or (lon, lat), or SpatialPoints)

sp2
coordinates for species 2 (matrix with (x, y) or (lon, lat), or SpatialPoints)

predictors
Raster object with environmental variables

n
integer. Number of randomizations

model
function. modeling algorithm to me used

verbose
logical. If TRUE some progress indicators are printed

...
additional arguments (none)

Value

numeric

Author(s)

Brian Anacker. Based on a similar function in by Christoph Heibl in the phyloclim package

References

Warren, D.L., R.E. Glor, M. Turelli, 2008. Environmental niche equivalency versus conservatism: quantitative approaches to niche evolution. Evolution 62:2868-2883.

[Package dismo version 1.0-15 Index]