pairs {dismo}	R Documentation

Pair plots

Description

Pair plots of presence and absence (background) data.

Methods

pairs(x, v=NULL, pa='pa', hist=TRUE, cor=TRUE)

x 	 Object of class DistModel or derived from that class (such as 'MaxEnt', 'Bioclim')

v 	 numeric, to select a subset of pairs, e.g. v=1:3 to plot only the first three variables

pa 	 Character. Either 'pa', 'p', or 'a' to show presence and absence, presence, or absence data respectively

hist 	 logical. If TRUE a histogram of the values is shown on the diagonal

cor 	 logical. If TRUE the correlation coefficient is shown in the upper panels

Author(s)

Robert J. Hijmans

See Also

density, plot

[Package dismo version 1.1-1 Index]