pointValues {dismo}	R Documentation

point values

Description

Extract values from a Raster* object for point locations. This function adds a few options that can be useful in the context of species distribution modeling to extract function in the raster package.

Usage

pointValues(x, p, a, uniquecells = TRUE, na.rm = TRUE)

Arguments

x
A Raster* object

p
Points. Two-colum matrix or data.frame; or a SpatialPoints* object

a
Additional points (absences). Two-colum matrix or data.frame; or a SpatialPoints* object

uniquecells
Logical. If TRUE, each cell can be included only once (i.e. 'duplicate' points are removed)

na.rm
Logical. If TRUE, cell values of NA are not returned

Value

matrix

Author(s)

Robert J. Hijmans

See Also

extract

[Package dismo version 1.1-1 Index]