prepareData {dismo}	R Documentation

Prepare data for model fitting

Description

Simple helper function to prepare data for model fitting

Usage

prepareData(x, p, b, factors, xy=FALSE) 

Arguments

x
Raster* object

p
presence points

b
background (or absence) points

factors
vectors indicating which variables are factors (using layer names or numbers)

xy
logical. If TRUE, the first two columns of the returned data.frame will be the coordinates of p and b (labeled 'x' and 'y')

Value

data.frame with nlayers(x)+1 columns and nrow(p) + nrow(b) rows. The first column, 'pb' indicates whether a record represents presence '1' or background '0' values. The other columns have the values from the Raster* object.

Author(s)

Robert J. Hijmans

[Package dismo version 1.0-15 Index]