randomPoints {dismo}	R Documentation

Random points

Description

Generate random points that can be used to extract background values ("random-absence"). The points are sampled (without replacement) from the cells that are not 'NA' in raster 'mask'.

If the coordinate reference system (of mask) is longitude/latitude, sampling is weighted by the size of the cells. That is, because cells close to the equator are larger than cells closer to the poles, equatorial cells have a higher probability of being selected.

Usage

randomPoints(mask, n, p, ext=NULL, extf=1.1, excludep=TRUE, prob=FALSE, 
             cellnumbers=FALSE, tryf=3, warn=2, lonlatCorrection=TRUE)

Arguments

mask
Raster* object. If the object has cell values, cells with NA are excluded (of the first layer of the object if there are multiple layers)

n
integer. Number of points

p
Presence points (if provided, random points won't be in the same cells (as defined by mask)

ext
Extent object. Can be used to restrict sampling to a spatial extent

extf
numeric. Multiplyer to adjust the size of extent 'ext'. The default increases of 1.1 increases the extent a little (5% at each side of the extent)

excludep
logical. If TRUE, presence points are exluded from background

prob
logical. If TRUE the values in mask are interpreted as probability weights (and the values should be positive numbers (or NA)). NOTE: this currently only works for rasters of a relatively modest size (that can be loaded into RAM)

cellnumbers
logical. If TRUE, cell numbers for ]codemask are returned rather than coordinates

tryf
numeric > 1. Multiplyer used for initial sample size from which the requested sample size is extracted after removing NA points (outside of mask)

warn
integer. 2 or higher gives most warnings. 0 or lower gives no warnings if sample size n is not reached

lonlatCorrection
logical. If TRUE then correct for the fact that longitude/latitude is not a planar coordinate system

Value

matrix with coordinates, or, if cellnumbers=TRUE, a vector with cell numbers.

Author(s)

Robert J. Hijmans

[Package dismo version 1.0-15 Index]