response {dismo}	R Documentation

response plots

Description

Generate 'response plots', i.e. single variable response curves for a model

Usage

response(x, ...) 

Arguments

x
Model object that inherits from 'DistModel', e.g. 'MaxEnt'. Also works for some other models (e.g. GLM)

...
Additional arguments. See Details

Details

var 	 Variable to be plotted (if NULL, all variables will be plotted)

at 	 Function to indicate at what level the other variables should be. E.g. median (the default), mean, min, max. Note that currently the function fails when 'mean' is used and one of the variables is a factor. Instead of a function you can also supply a single-row matrix with values for each of the variables. If NULL, partial response plots as are computed (this can take a lot of time). Partial response plots show, for each variable, the average prediction for all data points that were used to fit the model (or other data, if supplied), for 10 regularly spaced values of the variable.

range 	 'pa' (default) or 'p'. Show responses for the range of the presence data (p) or presence and absence (background) data (pa) for models that use absence data.

expand 	 percentage to expand the range of values with. Default is 10

rug 	 Logical. If TRUE (the default) a 'rug' of deciles is plotted on the horizontal axes)

data 	 data.frame with data to use, normally this is the data used to fit the model and does not needs to be supplied as they are available from the model object

fun 	 predict function. The default is "predict"; but you may change this to e.g., function(x, y, ...) predict(x, y, type='response', ...).

... 	 Additional graphical parameters

Value

Used for the side-effect of a plot

Author(s)

Robert J. Hijmans

See Also

density, plot, pairs

[Package dismo version 1.1-1 Index]