voronoi {dismo}	R Documentation

Voronoi polygons

Description

Create Voronoi polygons for a set of points. (These are also known Thiessen polygons, and Nearest Neighbor polygons; and the technique used is referred to as Delauny triangulation.)

Usage

voronoi(xy, ext, eps=1e-09, ...)

Arguments

xy
SpatialPoints* or two column matrix with x and y coordinates

ext
Extent. Can be used to set the corners of the rectangular window enclosing the triangulation. The default is the data range plus 10 percent. See deldir

eps
Numerical tolerance used in triangulation. See deldir

...
Additional arguments (none)

Value

SpatialPolygonsDataFrame

Author(s)

This is based on the link[deldir]{deldir} function by Rolf Turner and code by Carson Farmer: http://www.carsonfarmer.com/?p=455

Examples

# points
p <- matrix(c(17, 42, 85, 70, 19, 53, 26, 84, 84, 46, 48, 85, 4, 95, 48, 54, 66, 74, 50, 48, 
      28, 73, 38, 56, 43, 29, 63, 22, 46, 45, 7, 60, 46, 34, 14, 51, 70, 31, 39, 26), ncol=2)
	  
v <- voronoi(p)
v

[Package dismo version 1.1-1 Index]